requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
anyio==4.12.1
asttokens==3.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
comm==0.2.3
cryptography==46.0.2
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
debugpy==1.8.20
decorator==5.2.1
dill==0.4.0
executing==2.2.1
filelock==3.25.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
ntlm-auth==1.5.0
numpy==1.23.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.2
ply==3.11
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-sat==1.8.dev23
PyYAML==6.0.3
pyzmq==27.1.0
qiskit==0.43.3
qiskit-aer==0.12.2
qiskit-ibmq-provider==0.20.2
qiskit-terra==0.24.2
regex==2026.2.28
requests==2.32.5
requests-ntlm==1.1.0
rich==14.3.3
rustworkx==0.17.1
safetensors==0.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
stevedore==5.5.0
symengine==0.9.2
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.6.0
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
